+++ config.h
@@ -137,6 +137,9 @@ Experimental - testing required and homing needs to be worked out.
 */
 #if !defined DELTA_ROBOT || defined __DOXYGEN__
 #define DELTA_ROBOT Off
+#if !defined MINIMUM_FEED_RATE
+#define MINIMUM_FEED_RATE 0.1f // (radians/min)
+#endif
 #endif
 
 /*! \def POLAR_ROBOT